Move settings_changed from NCB to Network Spoke (#1270354)
Based on the Radek's comment:
I am thinking of NetworkControlBox as a class/widget that could be
eventually re-usable on its own (ie not embedded in the spoke). I have
some patches to make it living in its own module but have still troubles
splitting glade file and feeding only NCB's part to builder. So from
this point of view, I'd rather see settings_changed as an attribute of
the spoke (ie self.spoke in NCB).

Related: rhbz#1270354