Fix bad addon handling when addon import failed
An addon is added to AddonRegister even when the import of addon
failed. This happens because pykickstart is caling method
headerHandler() which thinks that addon wasn't add yet so the
placeholder is add instead. Because of the placeholder the setup() and
execute() methods of AddonData are called and in TUI install even the
"Executing {AddonName} addon" is printed.

Adding more warning messages when this happen and removing the
placeholder when setup() method is called.