Handle unmounting ostree when exiting
payload.unsetup() handles the unmounts so it needs to be called before
any image or dirinstall unmounting.

And some of the mounts are mounted multiple times, so if the unmount
fails log it and continue instead of crashing.

(cherry picked from commit b7785a125e1b10ac14f576ee03a26fc0b9393f96)