Don't join two absolute paths (#1249598)
The iutil.getSysroot() method returns an absolute path,
so we are effectively trying to join two absolute paths here.
The os.path.method() function just returns the first absolute path in
such case, which we certainly do not want.

So turn the path to the reconfigSys file to a relative one to
fix the issue.

Related: rhbz#1249598