Rerun writeBootLoader on Live BTRFS installs
In order to support installation of live media without an initramfs (to
save space) this reruns writeBootLoader for all live BTRFS
installations. This will also fix the missing initrd line when
/boot is on BTRFS (see rhbz#1200539).

Without this the initial grub2-mkconfig without an initramfs doesn't
build the kernel cmdline correctly. On live installations the initramfs
is always rebuilt, but that happens after the initial call to
writeBootLoader.