Move default X keyboard setting out of the Welcome spoke
Move the default X keyboard setting code from the Welcome
spoke to the keyboard module and trigger it from the Keyboard spoke.

This removes a dependency between the Welcome and Keyboard spokes,
that triggered a traceback if the Welcome spoke did not run but the
Keyboard spoke did.

Related: rhbz#1235726