Add support for device specification variants (#1200833)
Make it possible to have multiple device specification variants
delimited by | in the device specification string, for example:

"sd*|hd*|vda"

or

"sd*|hd*|vda,/dev/vdc"

The spec variants are processed individually from left to right
and at least one match needs to happen for the | delimited
spec to be considered successfully applied.

Also make sure that callers of the device_matches() function
really get disk devices when they expect them, not partitions
or other devices that are not disks.

And move the function to storage_utils.py - it really doesn't make
sense to have it in kickstart.py.

Resolves: rhbz#1200833