Remove the locales list from zanata.xml
Zanata stores this information on the server and no longer exports it as
part of the project config. Also, if the information in zanata.xml gets
out of sync with the server, po-pull crashes with a confusing 403 error.

To avoid the same problem with LINGUAS, remove LINGUAS from source
control and generate it based on the result of po-pull.

Related: rhbz#1304327