Deselect all addons correctly (#1333505)
If depsolve failed with some (or all) addons selected it wasn't cleared
by deselecting all of them because the package and group set were never
cleared.

Add a check to skip this if the addons have not been changed (or never
set, as is the case with kickstart).

(cherry picked from commit 2ad90e9294517770372f82a14792379a49bec6a2)