Ditch autopoint.
The build files provided by gettext are tailored to a specific type of
project, in which the translators are aware of the project layout and
code, and the translations are stored as part of the project. The first
assumption was never true, which is why the second no longer is.  That,
combined with gettext not really integrating with automake
(Makefile.in.in? really?) makes using the tools it provides a struggle.

Replace the autopoint Makefile with our own automake file. This file is
aware of srcdir vs. builddir so it can be used with distcheck without a
fight, and it requires no configuration files in the po directory.
Instead of POTFILES.in, look for translatable strings in every file.
Instead of LINGUAS, use whatever translations are present. The new
Makefile looks pretty big but it's like half comment; it's not that bad.

The po-empty target has been removed since it is no longer necessary. If
you don't want translations, just don't have translations. po-fallback
remains as non-failing po-pull.

The .pot file is no longer stored in git or added to the source tarball.
The .pot is generated from the source files, so do that at build time
instead, and before making a release so it can be pushed to Zanata.

Switch the glade file tests to skip tests of translated strings, since
that no longer works without LINGUAS and the results from it weren't
particularly helpful anyway.