Don't crash when selecting the same hdd ISO again (#1275771)
Don't crash if the same hdd ISO source is selected again.

The ISO selection mechanism unmounts the ISO file holding partition
after an ISO file is selected. This is usually not an issue as the payload
the will then mount it back once needed.

Unfortunately if the same ISO file was selected again, the attempt at unmounting
the partition would fail due to payload holding on to the iso file, resulting in a crash.

To avoid this we now always deselect the previous ISO before a new one is selected.

Resolves: rhbz#1275771