send all layouts to localed for keymap conversion (#1333998)
Resolves: rhbz#1333998

localed's code for deciding what VConsole keymap matches a given
X11 layout/variant config is more likely to work well when we
give it the whole list of X11 layout/variants we have, not just
the first. If you pass it only a single X11 layout/variant, it
will only work if there is an entry in the model map table that
is for *exactly* that layout/variant; for instance, if you pass
it 'ru', it will not consider its entry for 'ru,us' to be a
match. If you pass it a list of layouts, it's actually rather a
lot more likely to find a match - it would consider its 'ru,us'
entry to be a (weak) match for 'ru,foo,bar', for example,
because as a last resort it tries matching just the first part
(up to the first comma) of the layout against the first part of
the table entry, but *only* if you give it multiple layouts.