Stop using undocumented DNF logging API
DNF doesn't want users to access base.logging anymore.

This changes the dnf logging so that the dnf and dnf.rpm loggers go into
packaging.log now. It preserves the dnf.librepo.log file and the
hawkey.log files so that they can be examined after the installation is
complete or crashes.