Make the anaconda-generator exit early outside of the installation environment (#1289179)
Check if the anaconda-generator is running in the installation
environment by checking where the default.target is pointing.

Proceed normally if it is pointing to the anaconda.target,
exit early if not.

This should get rid of the annoying "anaconda-generator failed with
error code 1" errors in journal when the generator tried to run in
outside of the installation environment.

Resolves: rhbz#1289179