Fix some residual issues of dasdfmt in text-mode. (#1259437)
Some of these are dumb errors. For example, need to make sure we pass
a list of device names, not device objects, to the methods that determine
whether a DASD needs to be formatted.

devicetree.populate() needs to be called so that the devicetree is
actually made aware of the changes.

Make sure to call getDisks() when determining list of unformatted DASDs.
They won't be in data.ignoredisk.onlyuse since they are not entirely
"valid" disks due to their need to be formatted. Also can't just refer to
storage.devicetree.dasd since that list only contains DASDs that have
had dasdfmt run against them. So we need to call getDisks() against the
entire devicetree, where unformatted DASDs _do_ actually appear, and they
will be detected as disks since I added that logic in bfc1de46.

Additionally, only call storageInitialize and self._initialize() after
all of this if we're doing a manual installation. This makes sure things
like the summary screen get updated with accurate information about any
newly-formatted DASDs.

Resolves: rhbz#1259437