Fix the name sensitivity in the custom spoke.
When switching from an existing device to a non-existing device of the
same type, the sensitivity of the "Name" field was not being updated.
Add an extra call to the combo box change signal handler to handle this
case.

(cherry picked from commit ac832a104209acad425e706c57ebc81ae5716749)

Resolves: rhbz#1316269