Use the signal handlers to set initial widget sensitivies
Setting the initial sensitivities in the source spoke is a little
tricky, since the radio buttons determine the sensitivy for each box
corresponding to a method (and the children of that box), but the
protocol combo within one of those boxes also affects sensitivities.
And there's a case where got things wrong: if there is an auto-detected
repo, the auto-detect button would be active, but the network box would
still be sensitive.

Remove a bunch of the manual set_sensitive calls, and add signal handler
calls to refresh() to simulate changes to the protocol combo box as well
as a toggle for each radio button (buttons being disabled needs signals
too).

Resolves: rhbz#1259742