Fix adding new VG in Custom spoke can't be applied (#1263715)
When a new VG was created and nothing inside of a dialog was changed
then the changes can't be confirmed by the "Update Settings" button.

Method which runs the dialog assumed, when nothing was modified the
update button don't need to be updated. This method now gets info,
if this is a new VG or modification of an existing one.

Resolves: rhbz#1263715

Reported-by: Jan Stodola <jstodola@redhat.com>