Allow kickstart users to ignore the free space error
When attempting to install via kickstart with insufficient space for the
selected packages, let the user try to make it work anyway. For
graphical and text modes, the installation will stop at the summary hub
with the warning message but allow the user to begin the installation,
and in cmdline mode the installation will begin on its own.

In all cases the insufficient space message is written to anaconda.log.

Resolves: rhbz#1287308