Fix dhcpclass to work both via kickstart and the boot cmdline.
This does two things:

First, if inst.dhcpclass= was given on the command line, make sure the
/etc/dhclient.conf file is copied over to the anaconda environment.  NM
will read this file and merge its contents with the device-specific ones
before running dhclient.  This preserves the command line setting should
we re-run dhclient later (like, on a kickstart install).

Second, if network --dhcpclass= is given in a kickstart file, write out
a device-specific dhclient config blob in /etc/dhcp.  This file will
then get copied over to the anaconda environment.  NM will merge it,
just like in the first case.

Resolves: rhbz#1293051