Attempt to unload modules updated by a driver disk
For each module in a driver update disk, check if there is a module in
the initrd that matches the aliases provided by the new module, and
attempt to unload those existing modules before loading the updates.

Move the module loading and unloading so that occurs while the DUD is
not mounted. This way, if a module in the initrd was used to access the
DUD, it might still have a chance of being unloaded and updated by the
DUD.

(cherry picked from commit 7d7443b2959a25da13a706e5ab5c1e741b05abb6)