Support --when parameter in snapshot (#1113207)
--when can contain two values now:
"pre-install" = Do a snapshot of existing LV before installation.
                Beware the snapshot could be erased same as other
                partitions.
"post-install"= Do a snapshot after installation, as the last step of
                the installation.

Add warnings if clearpart and zerombr commands are used.
Commands clearpart and zerombr could destroy the pre-install snapshots
so warn user about this. This is not an error because it could be
useful to clear other partitions or origin partition without removing a
snapshot.

Related: rhbz#1113207