Fix restart payload only when repo needs network (#1358788)
When user change some settings in a network then the payload manager
will restart payload thread to test if repositories are still reachable.

Fix use-case when repositories are local so changing of network settings
don't require to test repositories.

Remove cdrom and hdd check because it's skipping the check on additional
repositories too. This check is contained in needsNetwork method already.

Resolves: rhbz#1358788

Reported-by: Peter Kotvan <pkotvan@redhat.com>