Improve connection network change detection
Reset payload when connection change.
This solve the issue only partially, problem here is that the
NetworkManager-connection-editor will save the connection in other form
than it is saved from the Dracut.

So when you click on the save button (without changing anything) for the
first time, the payload will be restarted but when you do it more times
after, it won't restart the payload thread anymore.

Based on RHEL 7.3 commit:
a58257430a14e664725b5ffa84eea21e8ef55

Doesn't work on master now because there is some issue with
NetworkManager connection activation.