Make it possible to override Zanata branch name
So far the makebumpver script made the assumption that
the name of the git branch and branch in the Zanata Anaconda
project are the same.

This might not necessarily be true anymore with the new Anaconda
branching scheme, so make it possible to both override the
branch name and to map known different names.