network: avoid registering GTypes from multiple threads (#1366271)
An attempt to derefrence an GEnum value causes a lazy registering of
GType which is not thread-safe. It is however done from the
wait_for_connecting_NM_thread() thread.

Let's just create the types that are potentially used from the
wait_for_connecting_NM_thread() thread in advance.

rvykydal@redhat.com: filed BZ for pygobject3: 1370124

Resolves: rhbz#1366271