network: adapt to changed NM ibft plugin enablement configuration (#1371188)
NM changed the way ibft plugin is configured to be anabled. Instead of special
NetworkManger/conf.d file configuring it as enabled, it is now enabled by
default. To disable the plugin in Anaconda (which should happen if there is no
ip=ibft boot option) we used to remove the respective config file. Now we drop
a file to disable the plugin.

Resolves: rhbz#1371188