Drop T from cp command flags when copying driver disk RPMs (#1269915)
We are copying the driver disk RPMs to a directory, but cp used the T
flag, which corresponds to:

-T, --no-target-directory
       treat DEST as a normal file

As the destination should always be a directory the T flag doesn't make
sense there. And indeed, if T is removed from the cp flags the driver disk
RPM is copied correctly.

Resolves: rhbz#1269915