Improved password quality checking
Refactor the password quality checking to make it more readable and
remove password checking code duplication in the user spoke, root password spoke
and in the LUKS passphrase dialog.

The checks have been moved to the GUI helpers module and are used
from all relevant places.

Also fixes a number of subtle password checking bugs and improves
the feedback messages provided to the user if a password is rejected
as bad.

Resolves: rhbz#1383718
Resolves: rhbz#1380277
Resolves: rhbz#1356975