rpmostreepayload: Fix remote handling to use correct sysroot
I'm working on Atomic Workstation, and putting the ostree remote
config into a package (same as CentOS is doing it), and this
is the current recommendation.

See
https://github.com/rhinstaller/anaconda/pull/474
for some previous discussion around this.

Recently, this broke; we end up deleting the remote in the package,
and the reason for this is we're using the *physical* sysroot rather
than the target, which means we look at the wrong
`/mnt/sysimage/etc/ostree/remotes.d`.

What I don't yet know is *why* this broke - I have a suspicion it's
related to some of the mount setup changes bcl did previously here,
but anyways, this patch fixes things.