Fix network spoke being incorrectly marked as mandatory (#1374864)
The network spoke should only be marked mandatory during an installation
from network sources.

So only mark it as mandatory when it runs in Anaconda & the installation
source indicates that it needs network.

Also use a modern way of checking if the installation source needs
network when we are at it.

Resolves: rhbz#1374864