Run efibootmgr for only one device in multipath (#1346725)
The efibootmgr tool was started for all devices in multipath and if
the last device failed then a whole installation will fail.

The issue should be fixed now by stopping the loop on first successful
run of efibootmgr tool. This is a multipath, therefore installing
bootloader to same device repeatedly is not a good idea.

Resolves: rhbz#1346725