Raise NonInteractive exception in Hubs event loop
Raise exception to stop installation when user interaction is required.

We ran event loop in every Hub (SummaryHub, ProgressHub) but we never
turned the event loop off so all the event loops processed the same
queue of events in parallel.
Now when a new Hub is shown, it will first disable old running event loops
and then it will enable its own loop.