Fix picking mountpoint for package download (#1328151)
Mountpoint where we are downloading packages was picked by the same
method as what is used in space check. This method is not working
correctly when user skipped space check, it still taking install size
in count.

When the first method failed, try again but don't include install size
and try to prefer other than root mountpoint for downloading.

Related: rhbz#1328151

Reported-by: Orion Poplawski <orion@cora.nwra.com>