iSCSI: adjust to change in blivet auth info (#1378156)
When blivet switched from using libiscsi to storaged, the node
attribute of an `iScsiDiskDevice` changed from being a libiscsi
`PyIscsiNode` instance to a blivet `NodeInfo` namedtuple, but
both blivet and anaconda continued trying to call the `getAuth`
method of the node - which `NodeInfo` doesn't have - to find
the authentication info for the node. I've sent a blivet PR
to store this information in the `NodeInfo` class instead:
https://github.com/rhinstaller/blivet/pull/518 . This patch
adjusts anaconda to use it.

Resolves: rhbz#1378156