Add git-find-branch script for finding parent branch
To add branch specific stuff, we need to find out on what main devel
branch (for example f25-devel, master, rhel7-branch) we are based on.
This script will find nearest parent by comparing nearest commit which
is on tested and actual branch.

Output of this script is name of the nearest base branch without the
`-*` suffix. This will handle f25-devel vs f25-release and it should
be easier for substitutions.

This won't work without a merge commit! Git doesn't have any information
which branch is the parent without the merge commit, both branches tips
are pointing to the same commit.