Add install task processing classes and unit tests
Add the Task and TaskQueue classes aimed on structured
sequential installation task processing.

The general idea is that individual installation tasks are wrapped by
the Task class and added to TaskQueues in the required order of execution.
TaskQueues can contain not only Tasks, but also other TaskQueues, this
enabling nesting.

TaskQueue simulates a list and behaves accordingly. Most list methods
are supported but list methods reordering and extending the TaskQueue
with another one are intentionally left unimplemented.

Once started the TaskQueue iterates over all Tasks and TaskQueues it contains
and starts them sequentially. Only one task is running at the
same time and there is no built-in parallelism in the TaskQueue.
The general expectation is that installation specific tasks need
to happen sequentially and any parallelism will be internal to individual
tasks.

Also note that both the Task and TaskQueue can be started only once -
additional calls to the start() method are ignored and only result in
an error being logged. This reflects the nature of installation
tasks which should usually be executed only once & where repeated
unintended execution will usually result in an error or
even a harmful effect.

The Task class provides started & completed signals. The TaskQueue
also has these two signals, but adds queue_started/queue_completed and
task_started/task_completed signals, which are triggered every time
a queues/task contained in the given TaskQueue is started/completed.

Callback for all these signals are given the Task/TaskQueue instance
that triggered the signal and first argument.

There are also unit tests for task Task and TaskQueue classes.