Move Anaconda tests to mock
To run tests in mock use `make isolated-ci`.

This is using copr repository from vtrefny so our packages will be
one day old from this repo and will fallback on Rawhide version when
Rawhide has newer packages (broken builds on COPR for example).

To run these tests the nosync package is required, it is for disabling fsync
in the mock and it will give us great performance boost. When you are
on a multilib (x86_64) system you need to install x86_64 and i686 versions
of the nosync library.