Add a signal/slot implementation
The Signal class provides a signal/slot implementation that
can be used to make code less tightly coupled and more modular.

Lets say that we have a class that needs to call a callback once
something happens. The class doesn't care about what the callback does,
if there is one or if there are multiple callbacks.

With this signal/slot implementation it simply instantiates the Signal
class and exports it as part of it's API:

self.task_complete = Signal()

Once the event of interest happens, it calls the emit method of the signal
(in this case with an argument):

self.task_complete.emit(result)

Other classes that care about the given event can connect their callbacks like this:

def _task_done_cb(self, task_result):
    log.info("Task %s completed." % task_result.name)

foo.task_complete.connect(self._task_done_cb)

Many callback can be connected to a single signal.
To disconnect a callback from a signal just pass the callback to the disconnect() method
and it will be disconnected from the signal:

foo_task_complete.disconnect(self._task_done_cb)

It is also possible to clear all callback from a signal with the clear() method:

foo.task_complete.clear()