ibft: fix dracut args and ifcfg values for ksdevice=bootif (non-ibft) (#1396949)
Resolves: rhbz#1396949

The use case is non-ibft device activated (eg pxe device with ksdevice=bootif)
and ibft devices on the same subnet. Pxe device is taking default route, ibft devices
are using the pxe device (default route) to reach iSCSI target and as a
consequence wrong dracut boot options (ip=<pxe device>) and ifcfg values
(NM_CONTROLLED=no is set not for ibft devices but for the device with route to
iscsi target, which is the pxe device) are written out.