RTD fixes
A couple of fixes to make the documentation look better on Read the Docs:
- add missing rst files to the index
- reorganize the index to move more important items to the top
- make sure the contribution guidelines are included
- add a header to contribution guidelines to make it show correctly on RTD

Also disable API docs generation on RTD - at least for now. It doesn't seem to
work very well at the moment and needs a big heap of dependencies (Pykickstart, Blivet)
not available on RTD to generate. We might be able to make it work in the future
by using an intermediate project to pre-build the API rst files.