network: don't reactivate devices from linuxrc.s390 for early networking (#1329171)
Resolves: rhbz#1329171

We shouldn't reactivate the devices (activated in s390 init linuxrc.s390) in
loader when trying to bring up early networking on s390.

For ipv6 only configuration the reactivating fails. The bug is hit when
installation (loader) is initiated by ssh -X install@<IP_ADDR> on s390, ie with
X forwarding turned on (-X). In this case loader assumes early networking is
required.

As a conservative fix the patch modifies the condition to leave early
networking setup if network was activated in s390 init (NM is connected
and network configuration from s390 init was found in loader).