Create lvm.conf with thin_* path overrides (#1350296)
Turns out loader remounts /sbin to /usr/sbin for stage2,
which effectively means we can't put anything to /sbin
through the upd-instroot script.

As LVM tryes to call the the thin_* tools from /sbin this is a problem.

To wor around this create /etc/lvm/lvm.conf file in the image that
overrides paths to the thin_* tools to point to /usr/sbin.

Resolves: rhbz#1350296