bootloader: Only install, don't write config for rpmostreepayload
The intention from the very start of RPMOSTreePayload is that it owns
writing the bootloader *configuration* (with some kernel args computed
by Anaconda), and Anaconda handles writing the MBR etc.  It basically
has to work this way because the fundamental basis of ostree is atomic
updates via swapping the bootloader config.

Except it never worked that way apparently - I think we'd just been
lucky until recently having the RPMOSTreePayload's config write
after the bootloader.

The Anaconda invocation of `grub2-mkconfig` recently blew up for some
reason I haven't debugged just on EFI, but since we don't want it
writing the configs, let's change the code to stop doing that.