Add a bottom bar to the Blivet GUI spoke
Add a bottom bar containing a summary button on the left
and and the "Undo last action" and "Reset all" buttons
on the right to the Blivet GUI spoke.

Also add a viewport Box called BlivetGuiViewport for holding the main
Blivet GUI widget.

Also add button for showing list of scheduled action to blivet-gui spoke