Validate dasd and zfcp user input (#1335092)
In gui, the user input for dasd and zfcp devices should be validated
before it is sanitized and the discovery is run. Otherwise, the
blockdev.s390.sanitize_* functions could cause the out of memory
error.

Also, the gui dialogs for dasd and zfcp were polished a little and
their behaviour made a little more consistent.

Resolves: rhbz#1335092