network: bind slave connections to DEVICE, not HWADDR (#1373360)
Related: rhbz#1373360

Bind to DEVICE for bond and bridge slaves same as we have been doing for team.
And do the same with connections created in Anaconda (ie with kickstart %pre).

Binding to HWADDR results in HWADDR="" for use case when kickstart is fetched
with ks=hd: or ks=file: and network is activated after parsing kickstart - so
when parsing it HWADDR info is not available in sysfs yet.