network: activate bridge for first network command in ks via %pre (#1416655)
Resolves: rhbz#1416655

Use case:
- ens3 activated in dracut
- kickstart command generated in %pre section:

%include /tmp/ksinclude
%pre
echo "network --device=bridge0  --bridgeslaves=ens3 --bootproto=dhcp" >> /tmp/ksinclude
%end

For the first network command in kickstart --activate is default.  If bridge
was configured in the command and its slave had already been activated in
dracut, the disconnection of slave which needs to happen before activating the
bridge (in _add_slave_connection) in anaconda did not happen because of the
value of networkdata.activate being None (special case for the first network
command). As a consequence the brige was not activated by anaconda.