Fix Anaconda forces payload restart when network (not)change (#1373449)
This problem is solved by downloading content of repomd.xml file and
comparing its MD5 hashes with downloaded repomd.xml file from the old
repo url.

MD5 hash is used here because it takes less memory and we don't need
repomd.xml content for anything else.

The postSetup() method was added to the payload. This method can be
overridden by Payload child classes which needs to do additional tasks
on the end of the restartThread call from the PayloadManager class.

Resolves: rhbz#1373449
Resolves: rhbz#1358778