Replace sanity check with more advanced storage checker
The function sanity_check was replaced with many small functions
for storage checking that can be registered with storage_checker
and called with defined constraints. This enables us to make the
storage checking more flexible, since we can dynamically register
new checks, change the constraints or just skip some checks.

SanityError and SanityWarning exceptions were removed, because they
were never raised and we only used them to distinguish between errors
and warnings. That is not necessary now.

Related: rhbz#1412151