Perform recursive copying of driver disk RPM repo contents
Commit anaconda-25.20-1-100-ge0e86f4 introduced a new way of copying RPM
repository data from a driver disk, which apparently is not recursive.
Since repository metadata is stored in a separate directory "repodata",
this now produces a warning like this:

    DD: WARNING: DD repo content not a file: /media/DD-1/rpms/x86_64/repodata

and skips repodata directory entirely. Luckily, it is generated later by
the code in pyanaconda.packaging,PackagePayload.addDriverRepos().

Unfortunately, test case was also written in assumption that repodata is
just a file and not a directory, so it hasn't caught it.

Resolves: rhbz#1425760

* dracut/driver_updates.py (save_repo): Implement recursive copying.
* tests/dracut_tests/test_driver_updates.py (makerepodata): New function
for populating fake repo metadata.
(TestSaveRepo.test_folder_repo): Update test in order to check for the
case of repodata being a directory.